initQuestion in case that the flavor is runned locally or as a fork
