+++ packages/cycle-scripts/scripts/init.js
@@ -4,8 +4,7 @@ const fs = require('fs-extra')
 const path = require('path')
 const chalk = require('chalk')
 const spawn = require('cross-spawn')
-const inquirer = require('inquirer')
-
+const initQuestions = require('./initQuestions')
 // Ask the user for which language and which stream library he want to use
 const dependencies = {
   basics: [
@@ -48,36 +47,6 @@ const replacements = {
     stream: 'most'
   }
 }
-
-const initQuestions = [
-  {
-    type: 'list',
-    name: 'language',
-    default: 0,
-    choices: ['JavaScript', 'TypeScript'],
-    message: 'Which language do you want to use to write your cycle app?'
-  },
-  {
-    type: 'list',
-    name: 'streamLib',
-    default: 0,
-    choices: [
-      {
-        name: 'XStream, tailored for Cycle.js',
-        value: 'xstream'
-      },
-      {
-        name: 'Most.js, a blazing fast stream library',
-        value: 'most'
-      },
-      {
-        name: 'RxJS',
-        value: 'rxjs'
-      }
-    ],
-    message: 'Which reactive stream library do you want to use?'
-  }
-]
 
 // function patchGitignore (appPath) {
 //   // Rename gitignore after the fact to prevent npm from renaming it to .npmignore
@@ -129,91 +98,96 @@ function successMsg (appName, appPath) {
   console.log()
 }
 
-module.exports = function init (appPath, appName, verbose, originalDirectory) {
+function setup (appPath, appName, verbose, originalDirectory, options) {
   const ownPackageName = require(path.join(__dirname, '..', 'package.json')).name
   const ownPath = path.join(appPath, 'node_modules', ownPackageName)
   const appPackageJson = path.join(appPath, 'package.json')
   const appPackage = require(appPackageJson)
 
-  inquirer.prompt(initQuestions).then(answers => {
-    const language = answers.language
-    const streamLib = answers.streamLib
+  const language = options.language
+  const streamLib = options.streamLib
 
-    const basicDependencies = dependencies.basics
-    const languageDependencies = dependencies.language[language]
-    const streamLibDependencies = dependencies.streamLib[streamLib]
+  const basicDependencies = dependencies.basics
+  const languageDependencies = dependencies.language[language]
+  const streamLibDependencies = dependencies.streamLib[streamLib]
 
-    const depsToInstall = basicDependencies.concat(languageDependencies).concat(streamLibDependencies)
+  const depsToInstall = basicDependencies.concat(languageDependencies).concat(streamLibDependencies)
 
-    // Manipulate app's package.json
-    // To be moved to separate module
-    appPackage.dependencies = appPackage.dependencies || {}
-    appPackage.devDependencies = appPackage.devDependencies || {}
-    appPackage.scripts = {
-      'start': 'cycle-scripts start',
-      'test': 'cycle-scripts test',
-      'build': 'cycle-scripts build',
-      'eject': 'cycle-scripts eject'
+   // Manipulate app's package.json
+  // To be moved to separate module
+  appPackage.dependencies = appPackage.dependencies || {}
+  appPackage.devDependencies = appPackage.devDependencies || {}
+  appPackage.scripts = {
+    'start': 'cycle-scripts start',
+    'test': 'cycle-scripts test',
+    'build': 'cycle-scripts build',
+    'eject': 'cycle-scripts eject'
+  }
+
+  fs.writeFileSync(
+    appPackageJson,
+    JSON.stringify(appPackage, null, 2)
+  )
+
+  // Copy flavor files
+  // fs.copySync(path.join(ownPath, 'template'), appPath)
+
+  fs.ensureDirSync(path.join(appPath, 'public'))
+  fs.copySync(path.join(ownPath, 'template/public'), path.join(appPath, 'public'))
+
+  // copy src and transform each of the file
+  fs.ensureDirSync(path.join(appPath, 'src'))
+  const templatePath = path.join(ownPath, 'template/src', language)
+  fs.readdir(templatePath, (err, files) => {
+    if (err) {
+      throw err
     }
-
-    fs.writeFileSync(
-      appPackageJson,
-      JSON.stringify(appPackage, null, 2)
-    )
-
-    // Copy flavor files
-    // fs.copySync(path.join(ownPath, 'template'), appPath)
-
-    fs.ensureDirSync(path.join(appPath, 'public'))
-    fs.copySync(path.join(ownPath, 'template/public'), path.join(appPath, 'public'))
-
-    // copy src and transform each of the file
-    fs.ensureDirSync(path.join(appPath, 'src'))
-    const templatePath = path.join(ownPath, 'template/src', language)
-    fs.readdir(templatePath, (err, files) => {
-      if (err) {
-        throw err
-      }
-      files.forEach(file => {
-        const targetPath = path.join(appPath, 'src', file)
-        const fileSrc = require(path.join(templatePath, file))
-        const targetSrc = fileSrc(replacements[streamLib])
-        fs.outputFile(targetPath, targetSrc)
-      })
-    })
-
-    fs.copySync(path.join(ownPath, 'template/src', language), path.join(appPath, 'src'))
-
-    // for each file in template/src load them, replace and write them
-
-    // for each library
-
-    // patchGitignore(appPath)
-
-    const dependecyList = depsToInstall
-      .slice(0, (depsToInstall.length - 1))
-      .join(', ')
-      .concat(` and ${depsToInstall.slice(-1)}`)
-
-    console.log(`Installing ${dependecyList} using npm...`)
-    console.log()
-
-    const args = [
-      'install'
-    ].concat(
-      depsToInstall
-    ).concat([
-      '--save',
-      verbose && '--verbose'
-    ]).filter(Boolean)
-
-    var proc = spawn('npm', args, {stdio: 'inherit'})
-    proc.on('close', function (code) {
-      if (code !== 0) {
-        console.error(chalk.red('`npm ' + args.join(' ') + '` failed'))
-        return
-      }
-      successMsg(appName, appPath)
+    files.forEach(file => {
+      const targetPath = path.join(appPath, 'src', file)
+      const fileSrc = require(path.join(templatePath, file))
+      const targetSrc = fileSrc(replacements[streamLib])
+      fs.outputFile(targetPath, targetSrc)
     })
   })
+
+  fs.copySync(path.join(ownPath, 'template/src', language), path.join(appPath, 'src'))
+
+  // for each file in template/src load them, replace and write them
+
+  // for each library
+
+  // patchGitignore(appPath)
+
+  const dependecyList = depsToInstall
+    .slice(0, (depsToInstall.length - 1))
+    .join(', ')
+    .concat(` and ${depsToInstall.slice(-1)}`)
+
+  console.log(`Installing ${dependecyList} using npm...`)
+  console.log()
+
+  const args = [
+    'install'
+  ].concat(
+    depsToInstall
+  ).concat([
+    '--save',
+    verbose && '--verbose'
+  ]).filter(Boolean)
+
+  var proc = spawn('npm', args, {stdio: 'inherit'})
+  proc.on('close', function (code) {
+    if (code !== 0) {
+      console.error(chalk.red('`npm ' + args.join(' ') + '` failed'))
+      return
+    }
+    successMsg(appName, appPath)
+  })
+}
+
+module.exports = function init (appPath, appName, verbose, originalDirectory, options) {
+  if (options) {
+    return setup(appPath, appName, verbose, originalDirectory, options)
+  }
+  return initQuestions(anwers => setup(appPath, appName, verbose, originalDirectory, anwers))
 }
